Avoiding cache overflow
In the controller:

I'm using an object literal to store the previous searches.
This is inefficient because we check the last updated time,
so new results will always appear forward (not backwards) in
time. This is as it should be. The matter is that I haven't
thought that the object literal is storing unneeded results,
meaning results that are too old to even be considered useful
for our purpose.

For that, now I'm storing results in an exclusive variable per
api (twitter, github), that storage is pour off every time
we get more results, right after we compare the stored results
with the new ones. After cleaned, the new results are stored
in that location, awayting for new results to come.